Yes made some changes
added some new topics

+++ src/main/java/com/example/demo/ServiceImpl/CategoryServiceImpl.java
@@ -9,11 +9,13 @@ import com.example.demo.DTO_and_ENUM.CategoryDto;
 import com.example.demo.DTO_and_ENUM.ProductDto;
 import com.example.demo.Dao.AdminDao;
 import com.example.demo.Dao.CategoryDao;
+import com.example.demo.Dao.LogInDao;
 import com.example.demo.Dao.ProductsDao;
 import com.example.demo.Exceptions.AdminException;
 import com.example.demo.Exceptions.CategoryException;
 import com.example.demo.Model.Admin;
 import com.example.demo.Model.Category;
+import com.example.demo.Model.LogIn;
 import com.example.demo.Model.Products;
 import com.example.demo.Service.CategoryService;
 
@@ -29,16 +31,26 @@ public class CategoryServiceImpl implements CategoryService{
 
 	@Autowired
 	private ProductsDao productsDao;
+	
+	@Autowired
+	private LogInDao logInDao;
+	
+	
 	@Override
-	public Category addCategory(CategoryDto categoryDto,Integer Id) throws CategoryException {
+	public Category addCategory(CategoryDto categoryDto,String uuid) throws CategoryException {
 	
 		Category c1= categoryDao.findByName(categoryDto.getName());
-		
-		Admin A1= admindao.findById(Id).orElseThrow(()-> new CategoryException("Admin Not Avail With the id -> "+Id));
-		
+
 		if(c1!=null) {
 			throw new CategoryException("Category Already Exist with the name "+categoryDto.getName());
 		}
+		
+		LogIn login = logInDao.findByUuid(uuid).orElseThrow(()-> new CategoryException("your not a valid admin"));
+		
+		Admin A1= admindao.findByUserName(login.getUsername());
+		
+		if(A1==null)
+				new CategoryException("Admin Not Avail With the id -> "+A1.getUserName());
 		
 		Category category= new Category(categoryDto.getName(),categoryDto.getDescription(), categoryDto.getImage1(),A1);
 		
@@ -57,10 +69,19 @@ public class CategoryServiceImpl implements CategoryService{
 	}
 
 	@Override
-	public Category updateCategory(Category category,Integer AdminId) throws CategoryException {
+	public Category updateCategory(Category category,String uuid) throws CategoryException {
+		
+		LogIn login = logInDao.findByUuid(uuid).orElseThrow(()-> new CategoryException("your not a valid admin"));
+		
+		Admin A1= admindao.findByUserName(login.getUsername());
+		
+		if(A1==null)
+			throw	new CategoryException("Admin Not Avail With the id -> "+A1.getUserName());
+		
 		
 		Category c1= categoryDao.findById(category.getId()).orElseThrow(()-> new CategoryException("Category Not Foound with the id"));
-		
+		if(A1.getCategory().equals(c1))
+			throw new CategoryException("your not a valid admin");
 		c1.setName(category.getName());
 		c1.setDescription(category.getDescription());
 		c1.setImage1(category.getImage1());
@@ -70,11 +91,20 @@ public class CategoryServiceImpl implements CategoryService{
 	}
 
 	@Override
-	public String deleteCategory(Integer id,Integer AdminID) throws CategoryException, AdminException {
+	public String deleteCategory(Integer id,String uuid) throws CategoryException, AdminException {
+		
+		LogIn login = logInDao.findByUuid(uuid).orElseThrow(()-> new CategoryException("your not a valid admin"));
+		
+		Admin A1= admindao.findByUserName(login.getUsername());
+		
+		if(A1==null)
+			throw	new CategoryException("Admin Not Avail");
 		
 		Category c1= categoryDao.findById(id).orElseThrow(()-> new CategoryException("Category Not Foound with the id"));
 		
-		Admin admin1=	admindao.findById(AdminID).orElseThrow(()-> new AdminException("You Dont have a Permission"));
+		if(A1.getCategory().equals(c1))
+			throw new CategoryException("your not a valid admin");
+		
 		
 		categoryDao.delete(c1);
